Added ModelStateTestBuilder.cs (#125)

+++ src/MyTested.Mvc.Core/IDefaultRegistrationPlugin.cs
@@ -5,7 +5,7 @@
 
     public interface IDefaultRegistrationPlugin
     {
-        int Priority { get; }
+        long Priority { get; }
         
         Action<IServiceCollection> DefaultServiceRegistrationDelegate { get; }
     }
